Check if profiler is initialized with a macro.
It looks like `collection` global variable is used to check if flask-profiler
is initialized or not. This prohibits an application from flexibly switching
environment from `DEBUG` to non-`DEBUG` mode.
This fix includes a macro that checks if `CONF` global is set,
as `CONF` is guaranteed to be set if flask-profile is initialized regardless of `DEBUG` mode.
This way, when `DEBUG` mode is turned off, code base does not need to be
modified.