Add production usage support by configurable sampling function
On high traffic websites, it is not ideal to sample each request via flask-profiler.
This commit provides an option to send a sampling function (provided by user)  as configuration parameter.

If the sampling function returns True, sample is taken, if sampling function is not defined or returns False, sample is not taken.

README.md is also edited in this commit to give a couple of usage examples.